fix: mechanism contact screen to contact update screen
