--- libs/ui/src/containers/screens/contact-form-screen/index.ts
@@ -1,2 +1,0 @@
-export * from './ContactFormScreen';
-export * from './ContactFormScreen.fetcher';
