+++ libs/ui/src/containers/screens/contact-update-screen/index.ts
@@ -1,0 +1,2 @@
+export * from './ContactUpdateScreen';
+export * from './ContactUpdateScreen.fetcher';
